requirements.txt
attrs==17.4.0
click==6.7
coverage==4.5.1
Flask==1.0
Flask-PyMongo==0.5.1
gunicorn==19.8.0
itsdangerous==0.24
Jinja2==2.10
#MarkupSafe==1.0
MarkupSafe==1.1.1
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pymongo==3.6.1
pytest==3.5.1
python-dotenv==0.8.2
six==1.11.0
Werkzeug==0.14.1
Flask-Migrate==2.2.1
#Flask-SQLAlchemy==2.3.2
Flask-SQLAlchemy==2.5.1
Flask-Babel==0.12.2
pycryptodome==3.7.3
PyMySQL==0.9.3
awscli==1.16.146
boto3==1.9.136
simplejson==3.16.0
Flask-SocketIO==4.2.1
python-socketio==4.3.0
python-engineio==3.9.0
gevent==1.4.0
SQLAlchemy<1.4